+++ infrastructure/web/templates/nomenclature/import_export.html
@@ -1,0 +1,25 @@
+{% extends 'base.html' %}
+
+{% block content %}
+    <div class="card w-50" style="background-color: transparent !important; border: 0;">
+        <div class="card-body">
+            <h2 class="mb-4">Импортировать услуги</h2>
+            <form method="post" enctype="multipart/form-data" action="{% url 'nomenclature_import' orgID=1 %}">
+                {% csrf_token %}
+                <div class="mb-3">
+                    <select name="nomenclature_id" class="form-select border-secondary bg-transparent">
+                        {% for nomenclature in nomenclatures %}
+                            <option value="{{ nomenclature.id }}">{{ nomenclature.name }}</option>
+                        {% endfor %}
+                    </select>
+                </div>
+                <div class="mb-3">
+                    <input type="file" class="form-control border-secondary bg-transparent" accept=".xlsx, .xls, .csv"
+                           name="excel_file" required>
+                    <div class="form-text text-dark">Файл должен быть в расширении .xslx, .xls, .csv</div>
+                </div>
+                <button class="btn btn-primary" type="submit">Загрузить</button>
+            </form>
+        </div>
+    </div>  
+{% endblock %}
