+++ infrastructure/web/templates/nomenclature/list.html
@@ -1,0 +1,105 @@
+{% extends 'base.html' %}
+{% load static %}
+
+{% block content %}
+    <div class="row p-2">
+            <select name="nomenclature" class="form-select col border-secondary bg-transparent me-2" id="nomenclature">
+                {% for nomenclature in nomenclatures %}
+                    <option id="option" value="{{ nomenclature.pk }}">{{ nomenclature }}</option>
+                {% endfor %}
+            </select>
+            <select id="category" class="col me-2 form-select border-secondary bg-transparent" name="category"
+                    aria-label="Default select example">
+                <option selected value="">Выберите категорию</option>
+                {% for category in categories %}
+                    <option value="{{ category.0 }}">{{ category.0 }}</option>
+                {% endfor %}
+            </select>
+            <select id="mark" class="col form-select border-secondary bg-transparent" name="mark"
+                    aria-label="Default select example">
+                <option selected value="">Выберите марку</option>
+                {% for mark in marks %}
+                    <option value="{{ mark.0 }}">{{ mark.0 }}</option>
+                {% endfor %}
+            </select>
+            <input id="search" class="col-6 form-control mt-2 border-secondary bg-transparent" type="search"
+                   placeholder="Поиск"
+                   aria-label="Search">
+    </div>
+    <div class="d-flex justify-content-between">
+        <div>
+            <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#importModal">
+                Импорт
+            </button>
+        </div>
+        {% if error %}
+            <div class="alert alert-danger d-flex align-items-center pt-1 pb-1 pe-4 ps-4 m-0" role="alert">
+                <svg xmlns="http://www.w3.org/2000/svg" width="15" height="15" fill="currentColor" class="bi bi-exclamation-triangle-fill flex-shrink-0 me-2" viewBox="0 0 16 16" role="img" aria-label="Warning:">
+                    <path d="M8.982 1.566a1.13 1.13 0 0 0-1.96 0L.165 13.233c-.457.778.091 1.767.98 1.767h13.713c.889 0 1.438-.99.98-1.767L8.982 1.566zM8 5c.535 0 .954.462.9.995l-.35 3.507a.552.552 0 0 1-1.1 0L7.1 5.995A.905.905 0 0 1 8 5zm.002 6a1 1 0 1 1 0 2 1 1 0 0 1 0-2z"/>
+                </svg>
+                <div>
+                    {{ error }}
+                </div>
+            </div>
+        {% endif %}
+        <div class="d-flex justify-content-center">
+            <input type="number" class="form-control me-1 bg-transparent border-secondary" style="width: 80px;" value="100" id="limit">
+            <div class="btn-group align-items-center">
+                <button class="btn btn-secondary" id="back">&#8592;</button>
+                <input type="number" class="form-control bg-transparent border-secondary" style="width: 50px" id="page" value="1" disabled>
+                <button class="btn btn-secondary" id="next">&#8594;</button>
+            </div>
+        </div>
+   </div>
+    <table id="table" class="table table-striped mt-3">
+        <thead>
+        <tr>
+            <th scope="col">Название</th>
+            <th scope="col">Категория</th>
+            <th scope="col">Примечание</th>
+            <th scope="col">Марка</th>
+            <th scope="col">Цена</th>
+        </tr>
+        </thead>
+        <tbody id="body"></tbody>
+    </table>
+    <div style="position: absolute; left: 50%; top: 50%;" id="beforeTable"></div>
+    <div class="modal fade" id="importModal" tabindex="-1" aria-labelledby="importModalLabel" aria-hidden="true">
+        <div class="modal-dialog">
+          <div class="modal-content">
+            <div class="modal-header">
+              <h4 class="modal-title" id="importModalLabel">Импортировать услуги</h4>
+              <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+            </div>
+            <form method="post" enctype="multipart/form-data" action="{% url 'nomenclature_import' orgID=1 %}">
+                <div class="modal-body">
+                    <div class="card" style="background-color: transparent !important; border: 0;">
+                        <div class="card-body">
+                            {% csrf_token %}
+                            <div class="mb-3">
+                                <select name="nomenclature_id" class="form-select border-secondary bg-transparent">
+                                    {% for nomenclature in nomenclatures %}
+                                        <option value="{{ nomenclature.id }}">{{ nomenclature.name }}</option>
+                                    {% endfor %}
+                                </select>
+                            </div>
+                            <div class="mb-3">
+                                <input type="file" class="form-control border-secondary bg-transparent" accept=".xlsx, .xls, .csv"
+                                    name="excel_file" required>
+                                <div class="form-text text-dark">Файл должен быть в расширении .xslx, .xls, .csv</div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button class="btn btn-primary" type="submit">Загрузить</button>
+                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Закрыть</button>
+                </div>
+            </form>
+          </div>
+        </div>
+    </div>
+{% endblock %}
+{% block js_block %}
+    <script src="{% static 'js/services/nomenclature_services.js' %}"></script>
+{% endblock %}
