+++ infrastructure/web/templates/partials/sidebar.html
@@ -32,9 +32,8 @@
           <div class="collapse sidebar-ul" id="price-list-collapse" style="">
             <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
                 <li><a href="{% url 'nomenclature_create' orgID=1 %}" class="link-dark sidebar-link rounded">Добавить номенклатуру</a></li>
-              <li><a href="{% url 'service_list' orgID=1 %}" class="link-dark sidebar-link rounded">Список услуг</a></li>
-              <li><a href="#" class="link-dark sidebar-link rounded">Добавить услугу</a></li>
-              <li><a href="{% url 'service_import' orgID=1 %}" class="link-dark sidebar-link rounded">Импорт и экспорт</a></li>
+                <li><a href="{% url 'nomenclature_list' orgID=1 %}" class="link-dark sidebar-link rounded">Список услуг</a></li>
+                <li><a href="#" class="link-dark sidebar-link rounded">Добавить услугу</a></li>
             </ul>
           </div>
         </li>
