--- infrastructure/web/templates/service/list.html
@@ -1,56 +1,0 @@
-{% extends 'base.html' %}
-{% load static %}
-
-{% block content %}
-    <div>
-        <form class="row">
-            <select name="nomenclature" class="form-select col border-secondary bg-transparent me-2" id="nomenclature">
-                {% for nomenclature in nomenclatures %}
-                    <option id="option" value="{{ nomenclature.pk }}">{{ nomenclature }}</option>
-                {% endfor %}
-            </select>
-            <select id="category" class="col me-2 form-select border-secondary bg-transparent" name="category"
-                    aria-label="Default select example">
-                <option selected value="">Выберите категорию</option>
-                {% for category in categories %}
-                    <option value="{{ category.0 }}">{{ category.0 }}</option>
-                {% endfor %}
-            </select>
-            <select id="mark" class="col form-select border-secondary bg-transparent" name="mark"
-                    aria-label="Default select example">
-                <option selected value="">Выберите марку</option>
-                {% for mark in marks %}
-                    <option value="{{ mark.0 }}">{{ mark.0 }}</option>
-                {% endfor %}
-            </select>
-            <input id="search" class="col-6 form-control mt-2 border-secondary bg-transparent" type="search"
-                   placeholder="Поиск"
-                   aria-label="Search">
-        </form>
-        <div>
-        </div>
-    </div>
-    <div  style="margin-right: 150px" class="d-flex justify-content-center mt-4">
-       <input type="number" class="form-control me-1 bg-transparent border-secondary" style="width: 80px;" value="100" id="limit">
-        <div class="btn-group align-items-center">
-            <button class="btn btn-secondary" id="back">&#8592;</button>
-            <input type="number" class="form-control bg-transparent border-secondary" style="width: 50px" id="page" value="1" disabled>
-            <button class="btn btn-secondary" id="next">&#8594;</button>
-        </div>
-   </div>
-    <table id="table" class="table table-striped mt-3">
-        <thead>
-        <tr>
-            <th scope="col">Категория</th>
-            <th scope="col">Название</th>
-            <th scope="col">Примечание</th>
-            <th scope="col">Марка</th>
-            <th scope="col">Цена</th>
-        </tr>
-        </thead>
-        <tbody id="body"></tbody>
-    </table>
-{% endblock %}
-{% block js_block %}
-    <script src="{% static 'js/services/nomenclature_services.js' %}"></script>
-{% endblock %}
