+++ models/nomenclature/admin.py
@@ -6,7 +6,7 @@ from .models import Nomenclature
 @admin.register(Nomenclature)
 class NomenclatureAdmin(admin.ModelAdmin):
     """Номенклатура"""
-    list_display = ('id', 'name', 'organization', 'services')
+    list_display = ('id', 'name', 'organization')
     list_display_links = ('id', 'name',)
     fieldsets = (
         (None, {
