+++ requirements.txt
@@ -18,6 +18,7 @@ django-filter==21.1
 django-import-export==2.8.0
 djangorestframework==3.13.1
 et-xmlfile==1.1.0
+excel2json-3==0.1.6
 exceptiongroup==1.0.0rc7
 flake8==4.0.1
 future==0.18.2
@@ -33,9 +34,11 @@ MarkupPy==1.14
 MarkupSafe==2.1.1
 mccabe==0.6.1
 Naked==0.1.31
+numpy==1.22.4
 odfpy==1.4.1
 openpyxl==3.0.9
 packaging==21.3
+pandas==1.4.2
 platformdirs==2.5.2
 postgres==4.0
 prompt-toolkit==3.0.29
@@ -46,6 +49,7 @@ pyflakes==2.4.0
 pylint==2.13.9
 pyparsing==3.0.9
 pyrsistent==0.18.1
+python-dateutil==2.8.2
 pytz==2022.1
 PyYAML==6.0
 questionary==1.10.0
@@ -59,7 +63,7 @@ tablib==3.2.1
 termcolor==1.1.0
 tomli==2.0.1
 tomlkit==0.10.2
-typing-extensions==4.2.0
+typing_extensions==4.2.0
 tzdata==2022.1
 urllib3==1.26.9
 wcwidth==0.2.5
